Starting the conversion of DataArray to have both a tuple dimension and component dimension. Code does not compile. Not even remotely.

+++ Source/DREAM3DLib/Common/ModifiedLambertProjectionArray.h
@@ -78,17 +78,17 @@ class DREAM3DLib_EXPORT ModifiedLambertProjectionArray : public IDataArray
     DREAM3D_INSTANCE_PROPERTY(int, Phase)
     DREAM3D_INSTANCE_PROPERTY(QVector<ModifiedLambertProjection::Pointer>, ModifiedLambertProjectionArray)
 
-    virtual IDataArray::Pointer createNewArray(size_t numElements, int rank, int* dims, const QString& name)
+    virtual IDataArray::Pointer createNewArray(size_t numElements, int rank, size_t* dims, const QString& name)
     {
       return ModifiedLambertProjectionArray::New();
     }
 
-    virtual IDataArray::Pointer createNewArray(size_t numElements, std::vector<int> dims, const QString& name)
+    virtual IDataArray::Pointer createNewArray(size_t numElements, std::vector<size_t> dims, const QString& name)
     {
       return ModifiedLambertProjectionArray::New();
     }
 
-    virtual IDataArray::Pointer createNewArray(size_t numElements, QVector<int> dims, const QString& name)
+    virtual IDataArray::Pointer createNewArray(size_t numElements, QVector<size_t> dims, const QString& name)
     {
       return ModifiedLambertProjectionArray::New();
     }
